Second chance modifiction, clear pages controller
Second chance algorithm modification created. Might be improved later.
Currently it stores the list, list_begin(list) is the algorithm head is pointing to now.

When frame has to be saved somewhere, if it's created using PAL_ZERO and is not changed, we don't wate space for it.

Improved mmap(for dirty things)